feat: implement asset stock management; add stock increase and decrease endpoints, and integrate stock handling in asset service

+++ internal/services/assets/asset_image_service.go
@@ -5,8 +5,9 @@ import (
 	"asset-service/internal/models/assets"
 	repository "asset-service/internal/repository/assets"
 	"asset-service/internal/utils"
-	"encoding/json"
-	"github.com/nats-io/nats.go"
+	nt "asset-service/internal/utils/nats"
+	"errors"
+	"fmt"
 	"github.com/rs/zerolog/log"
 	"path/filepath"
 )
@@ -15,24 +16,23 @@ type AssetImageService interface {
 	AddAssetImage(assetRequest []response.AssetImageResponse, assetID uint, clientID string) error
 	GetAssetImageByAssetID(assetID uint) (*[]assets.AssetImage, error)
 	DeleteAssetImage(assetID uint, clientID string) error
-	Cleanup(nats string) error
-	CleanupUnusedImages(nats string) error
+	Cleanup() error
+	CleanupUnusedImages() error
 }
 
 type assetImageService struct {
 	AssetImageRepository repository.AssetImageRepository
 	AssetRepository      repository.AssetRepository
 	Redis                utils.RedisService
+	NatsService          nt.NatsService
 }
 
-func NewAssetImageService(
-	assetImageRepository repository.AssetImageRepository,
-	assetRepository repository.AssetRepository,
-	redis utils.RedisService) AssetImageService {
+func NewAssetImageService(assetImageRepository repository.AssetImageRepository, assetRepository repository.AssetRepository, redis utils.RedisService, natsService nt.NatsService) AssetImageService {
 	return &assetImageService{
 		AssetImageRepository: assetImageRepository,
 		AssetRepository:      assetRepository,
 		Redis:                redis,
+		NatsService:          natsService,
 	}
 }
 
@@ -98,43 +98,60 @@ func (s assetImageService) DeleteAssetImage(assetID uint, clientID string) error
 }
 
 // Cleanup removes images of deleted assets
-func (s assetImageService) Cleanup(nats string) error {
+func (s assetImageService) Cleanup() error {
 	deleted, err := s.AssetRepository.GetAssetDeleted()
 	if err != nil {
 		log.Error().Str("key", "GetAssetDeleted").Err(err).Msg("Failed to get deleted assets")
 		return err
 	}
 
+	if len(deleted) == 0 {
+		log.Info().Msg("✅ No deleted assets found, cleanup not needed.")
+		return nil
+	}
+
 	for _, asset := range deleted {
+		// Fetch images associated with the asset
 		assetImages, err := s.AssetImageRepository.GetAssetImageByAssetID(asset.AssetID)
 		if err != nil {
 			log.Error().Str("key", "GetAssetImageByAssetID").Uint("asset_id", asset.AssetID).Err(err).Msg("Failed to get asset images")
 			return err
 		}
 
+		// Collect image filenames
 		var images []string
 		for _, img := range *assetImages {
 			images = append(images, filepath.Base(img.ImageURL))
 		}
 
+		// **VALIDATION: If no images, return an error**
+		if len(images) == 0 {
+			errMsg := fmt.Sprintf("❌ No images found for asset ID %d, skipping deletion", asset.AssetID)
+			log.Warn().Str("client_id", asset.UserClientID).Msg(errMsg)
+			return errors.New(errMsg)
+		}
+
+		// Delete asset image metadata from database
 		if err = s.DeleteAssetImage(asset.AssetID, asset.UserClientID); err != nil {
 			log.Error().Str("key", "DeleteAssetImage").Uint("asset_id", asset.AssetID).Err(err).Msg("Failed to delete asset images")
 			return err
 		}
 
-		if len(images) == 0 {
-			continue
-		} else {
-			if err = requestImageDeletion(nats, asset.UserClientID, images); err != nil {
-				log.Error().Str("key", "RequestImageDeletion").Err(err).Msg("Failed to request image deletion")
-			}
+		// Log the number of images being deleted
+		log.Info().Str("client_id", asset.UserClientID).Msgf("🗑️ Images to be deleted: %d", len(images))
+
+		// Request physical file deletion via NATS
+		if err = s.NatsService.RequestImageDeletion(asset.UserClientID, images); err != nil {
+			log.Error().Str("key", "RequestImageDeletion").Err(err).Msg("❌ Failed to request image deletion via NATS")
+			return err
 		}
 	}
 
+	log.Info().Msg("✅ Cleanup process completed successfully.")
 	return nil
 }
 
-func (s assetImageService) CleanupUnusedImages(nats string) error {
+func (s assetImageService) CleanupUnusedImages() error {
 	images, err := s.AssetImageRepository.GetAssetImage()
 	if err != nil {
 		log.Error().Str("key", "GetUnusedImages").Err(err).Msg("Failed to get unused images")
@@ -164,33 +181,9 @@ func (s assetImageService) CleanupUnusedImages(nats string) error {
 		log.Info().Msg("No unused images found")
 		return nil
 	} else {
-		if err = requestImageUsage(nats, assetImages); err != nil {
+		if err = s.NatsService.RequestImageUsage(assetImages); err != nil {
 			log.Error().Str("key", "RequestImageDeletion").Err(err).Msg("Failed to request image deletion")
 		}
 	}
 	return nil
-}
-
-func requestImageDeletion(natsAPI string, clientID string, images []string) error {
-	nc, err := nats.Connect(natsAPI)
-	if err != nil {
-		return err
-	}
-	defer nc.Close()
-
-	data, _ := json.Marshal(assets.ImageDeleteRequest{ClientID: clientID, Images: images})
-
-	return nc.Publish("asset.image.delete", data)
-}
-
-func requestImageUsage(natsAPI string, images []assets.ImageDeleteRequest) error {
-	nc, err := nats.Connect(natsAPI)
-	if err != nil {
-		return err
-	}
-	defer nc.Close()
-
-	data, _ := json.Marshal(images)
-
-	return nc.Publish("asset.image.usage", data)
 }
